+++ internal/blockstm/types.go
@@ -69,8 +69,11 @@ type MultiStore interface {
 
 // MVStore is a value type agnostic interface for `MVData`, to keep `MVMemory` value type agnostic.
 type MVStore interface {
-	Delete(Key, TxnIndex)
-	WriteEstimate(Key, TxnIndex)
+	InitWithEstimates(TxnIndex, Locations)
+	ConvertWritesToEstimates(txn TxnIndex)
+	ClearEstimates(txn TxnIndex)
+	ConsolidateEmpty(txn TxnIndex)
+
 	ValidateReadSet(TxnIndex, *ReadSet) bool
 	SnapshotToStore(storetypes.Store)
 }
@@ -79,6 +82,6 @@ type MVStore interface {
 type MVView interface {
 	storetypes.Store
 
-	ApplyWriteSet(TxnVersion) Locations
+	ApplyWriteSet(TxnVersion) bool
 	ReadSet() *ReadSet
 }